Modularized Backend Routes
- signup
- login
- fetch profile
- update profile
